# Conflicts:
#	package.json

+++ app/application/torii-adapter.js
@@ -15,8 +15,10 @@ function pushTokenToStore(tokenPayload, store) {
   return store.push('token', {
     id: tokenPayload.id,
     accessToken: tokenPayload.access_token,
+    rawPayload: JSON.stringify(tokenPayload),
     links: {
-      user: tokenPayload._links.user.href
+      user: tokenPayload._links.user.href,
+      actor: tokenPayload._links.actor && tokenPayload._links.actor.href
     }
   });
 }
@@ -32,7 +34,8 @@ export default Ember.Object.extend({
     }).then((token) => {
       return Ember.RSVP.hash({
         token,
-        currentUser: token.get('user')
+        currentUser: token.get('user'),
+        currentActor: token.get('actor')
       });
     }).then((session) => {
       // Load role eagerly
@@ -74,15 +77,22 @@ export default Ember.Object.extend({
       `A token must be passed: session.close('aptible', /*token*/);`,
       !!token
     );
-    return ajax(config.authBaseUri+`/tokens/${token.get('id')}`, {
-      type: 'DELETE',
-      headers: {
-        'Authorization': 'Bearer ' + getAccessToken()
-      },
-      xhrFields: { withCredentials: true }
-    }).then(() => {
-      clearSession();
-    });
+
+    let promise;
+
+    if (!!token.noDelete) {
+      promise = new Ember.RSVP.Promise((resolve, _) => resolve());
+    } else {
+      promise = ajax(config.authBaseUri+`/tokens/${token.get('id')}`, {
+        type: 'DELETE',
+        headers: {
+          'Authorization': 'Bearer ' + getAccessToken()
+        },
+        xhrFields: { withCredentials: true }
+      });
+    }
+
+    return promise.then(() => { clearSession(); });
   },
 
   identifyToAnalytics(user) {
